+++ src/RcppExports.cpp
@@ -11,42 +11,42 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // qualpal_cpp_rgb
-Rcpp::List qualpal_cpp_rgb(int n, const Rcpp::NumericMatrix& rgb_in, const Rcpp::List& cvd_list);
-RcppExport SEXP _qualpalr_qualpal_cpp_rgb(SEXP nSEXP, SEXP rgb_inSEXP, SEXP cvd_listSEXP) {
+Rcpp::List qualpal_cpp_rgb(int n, const Rcpp::NumericMatrix& rgb_in, const Rcpp::List& options);
+RcppExport SEXP _qualpalr_qualpal_cpp_rgb(SEXP nSEXP, SEXP rgb_inSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< const Rcpp::NumericMatrix& >::type rgb_in(rgb_inSEXP);
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type cvd_list(cvd_listSEXP);
-    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_rgb(n, rgb_in, cvd_list));
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type options(optionsSEXP);
+    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_rgb(n, rgb_in, options));
     return rcpp_result_gen;
 END_RCPP
 }
 // qualpal_cpp_colorspace
-Rcpp::List qualpal_cpp_colorspace(int n, const Rcpp::List& hsl_colorspace, const int n_points, const Rcpp::List& cvd_list);
-RcppExport SEXP _qualpalr_qualpal_cpp_colorspace(SEXP nSEXP, SEXP hsl_colorspaceSEXP, SEXP n_pointsSEXP, SEXP cvd_listSEXP) {
+Rcpp::List qualpal_cpp_colorspace(int n, const Rcpp::List& hsl_colorspace, const int n_points, const Rcpp::List& options);
+RcppExport SEXP _qualpalr_qualpal_cpp_colorspace(SEXP nSEXP, SEXP hsl_colorspaceSEXP, SEXP n_pointsSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< const Rcpp::List& >::type hsl_colorspace(hsl_colorspaceSEXP);
     Rcpp::traits::input_parameter< const int >::type n_points(n_pointsSEXP);
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type cvd_list(cvd_listSEXP);
-    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_colorspace(n, hsl_colorspace, n_points, cvd_list));
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type options(optionsSEXP);
+    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_colorspace(n, hsl_colorspace, n_points, options));
     return rcpp_result_gen;
 END_RCPP
 }
 // qualpal_cpp_palette
-Rcpp::List qualpal_cpp_palette(int n, const std::string& palette, const Rcpp::List& cvd_list);
-RcppExport SEXP _qualpalr_qualpal_cpp_palette(SEXP nSEXP, SEXP paletteSEXP, SEXP cvd_listSEXP) {
+Rcpp::List qualpal_cpp_palette(int n, const std::string& palette, const Rcpp::List& options);
+RcppExport SEXP _qualpalr_qualpal_cpp_palette(SEXP nSEXP, SEXP paletteSEXP, SEXP optionsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< int >::type n(nSEXP);
     Rcpp::traits::input_parameter< const std::string& >::type palette(paletteSEXP);
-    Rcpp::traits::input_parameter< const Rcpp::List& >::type cvd_list(cvd_listSEXP);
-    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_palette(n, palette, cvd_list));
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type options(optionsSEXP);
+    rcpp_result_gen = Rcpp::wrap(qualpal_cpp_palette(n, palette, options));
     return rcpp_result_gen;
 END_RCPP
 }